# Optimize FileClient for latency vol2