### What does this PR do?

* Added new master function _file_stats which provides a list of files with stat(..) info for them
* Changed FileClient to avoid hashing remote and local files (helps a lot when you deal with binary files!)
* Changed FileClient to stop contacting the master for every requested file, if it's up-to-date based on cached last-modified dates, which are fetched only once with _file_stats
* Currently only the 'roots' filesystem is fully supported with this new feature and other filesystems like 'gitfs' will fallback to comparing hashes and using _file_hash_and_stat for every request

I tried really hard to avoid breaking any functionality I might personally not use, like gitfs or Windows support, and cleaned the get_file function a bit.
If I could have missed some really edge case please let me know 😄 
(Note: according to https://docs.python.org/2/library/os.html os.stat is supported without problems for windows with file size and modification time. We just need to ignore the chmod part)

This PR depends on #39412 - the fix of FileClient caching - to show real performance increase.
**Combination of this and #39412 reduced our deploy times today from 18 seconds to 12 seconds(!)
(master-minion latency - 10 ms ping) - sounds cool, huh?**

**Note: currently a combination of FileClient + file.managed is very buggy for big binary files, or at least it was 3 months ago when I last checked. (jars, tars....etc)
It was so bad I just decided that I don't care and made a formula to make a nice rsync server, and rsync the files.
I'll try to make support for binary files better to the point where it would be on-par with rsync and I could remove rsync server.
Surely removing the hashing here is a big improvement itself, because previously it could take several seconds just to check if some binary files is up-to-date
There were some bugs(big performance problems) in binary portion of file.managed itself and I'll try to get back to that bugs soon.**

### What issues does this PR fix or reference?

As a follow-up to my previous PR, there's still room for improvement in FileClient's latency that I aim to fix now.

### Previous Behavior
Call _file_hash_and_stat for every requested file (ping-pong with master), hash local and remote files even if they are up-to-date.

### New Behavior
Call _file_stats once and cache the result in FileClient.
Fail fast with "file not found" if file is not in cached list of files.
Up-to-date check will now compare local and remote modification dates as well as file size instead of using hashes.

If the file is up-to-date, master will not be queried for this file, thus the latency to check if it's up-to-date will now be 0ms.

### Tests written?

No

Please review [Salt's Contributing Guide](https://docs.saltstack.com/en/latest/topics/development/contributing.html) for best practices.