@lorengordon 2017-02-15
Would os.scandir() help here? os.stat() is pretty slow itself, no?

http://benhoyt.com/writings/scandir/

It's included in python 3.5, but the pypi library is compatible back to python 2.6.