@ninja- 2017-02-15
@lorengordon this isn't called too often, the file list is cached for some amount of time(?) I think using os.scandir() would break symlink support (which I personally don't use or care about)