@terminalmage 2017-02-15
@ninja- the cached file lists live for 20 seconds by default before being aged out.