@ninja- 2017-02-15
@terminalmage that's pretty reasonable, but is it expired when highstate is triggered or something? Otherwise the deploy scripts could break, but I never saw it break like that.

@cachedout 
![](http://i2.kym-cdn.com/entries/icons/facebook/000/010/809/happy-oh-stop-it-you.jpg)