@terminalmage 2017-02-15
@ninja- no, this is a master-side cache, and the whole reason these file list caches exist on the master is to fix a performance bottleneck when hundreds/thousands of minions start running a highstate. Walking the file_roots (or making GitPython/Pygit2) requests to get file lists separately for each minion was causing a big performance issue, hence the caching.

If you want to clear the file lists cache before triggering a highstate, there is a runner function for that: https://docs.saltstack.com/en/latest/ref/runners/all/salt.runners.fileserver.html#salt.runners.fileserver.clear_file_list_cache