@ninja- 2017-02-15
@terminalmage _file_find wasn't cached that way and _file_list is hm, I'll need to think on how to change it so it won't break deploy scripts with a race condition(returning FileNotFound for recently added file)...