@ninja- 2017-02-15
look, Linux changes directory's lastmodified when any of the files inside is modified, according to my quick test.

We could use that information to refresh the cache without performance problems.

Is salt-master on windows supported? Is that something I could care about?