@lorengordon 2017-02-15
> Is salt-master on windows supported? Is that something I could care about?

Not yet, I don't think. But masterless mode is. i.e. `C:\salt\salt-call.bat --local`.