@cachedout 2017-02-16
Yes, salt-master on Windows is something that will need to work for this release.