@ninja- 2017-02-16
The only business-logic thing left here is to purge the file list cache in the way I mentioned
(I saw myself that it's returning false up-to-date during deploys) and it will be ready for final review. I'll get back to that sooon.