@ninja- 2017-02-17
@cachedout it's a conflict with @terminamage's symlink fix. I'll rebase.