@ninja- 2017-03-21
@cachedout I left a comment here https://github.com/saltstack/salt/pull/39412#issuecomment-288238010 about the situation with this PR, Basically for this PR to make sense, I need to make a proper thread-local caching of FileClient instances and I am out of ideas how to do it by myself without breaking half of the codebase.