@cachedout 2017-03-28
Since this now has merge conflicts I am going to close it for the time being. When ready, we can re-open it if necessary. Thanks.